fix: service status check fix: containerStatusJob fix: service form
